pyproject.toml
[project]
name = "workspace-migration"
version = "0.1.0"
description = "DAB-packaged tool for migrating Unity Catalog data between Azure Databricks workspaces"
requires-python = ">=3.10"
dependencies = [
    "databricks-sdk>=0.30.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "ruff>=0.6.0",
    "mypy>=1.11",
]

[[tool.uv.index]]
url = "https://pypi-proxy.dev.databricks.com/simple"
default = true

[tool.ruff]
target-version = "py310"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
# Integration notebooks carry long inline SQL strings; splitting them to
# 120 cols hurts readability. Unit test_cross_cutting has long pattern
# strings for source-scan assertions.
"tests/integration/*.py" = ["E501", "E402"]
"tests/unit/test_cross_cutting.py" = ["E501"]

[tool.mypy]
python_version = "3.10"
mypy_path = "src"
explicit_package_bases = true
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "databricks.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pyspark.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "yaml"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "databricks-sql-connector>=3.0",
    "sqlglot>=25.0",
]
